Move Contacts to geometry module
